refactor: :recycle: decision createAttribute component for reuse
