+++ src/features/createAttribute/index.ts
@@ -1,0 +1,1 @@
+export { CreateAttribute } from './ui';
