+++ src/features/createAttribute/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as CreateAttribute } from './CreateAttribute';
